LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/api/basic/mode-same-origin.any-expected.txt

CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/fetch/api/resources/top.txt.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/top.txt.
CONSOLE MESSAGE: Unsafe attempt to load URL https://web-platform.test:9443/fetch/api/resources/top.txt?location=https%3A%2F%2Fweb-platform.test%3A9443%2Ffetch%2Fapi%2Fresources%2Ftop.txt&count=1 from origin http://web-platform.test:8800. Domains, protocols and ports must match.

CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/fetch/api/resources/top.txt?location=https%3A%2F%2Fweb-platform.test%3A9443%2Ffetch%2Fapi%2Fresources%2Ftop.txt&count=1 due to access control checks.
CONSOLE MESSAGE: Unsafe attempt to load URL http://www1.web-platform.test:8800/fetch/api/resources/top.txt?location=http%3A%2F%2Fwww1.web-platform.test%3A8800%2Ffetch%2Fapi%2Fresources%2Ftop.txt&count=1 from origin http://web-platform.test:8800. Domains, protocols and ports must match.

CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/top.txt?location=http%3A%2F%2Fwww1.web-platform.test%3A8800%2Ffetch%2Fapi%2Fresources%2Ftop.txt&count=1 due to access control checks.

PASS Fetch ../resources/top.txt with same-origin mode
PASS Fetch http://web-platform.test:8800/fetch/api/resources/top.txt with same-origin mode
PASS Fetch https://web-platform.test:9443/fetch/api/resources/top.txt with same-origin mode
PASS Fetch http://www1.web-platform.test:8800/fetch/api/resources/top.txt with same-origin mode
PASS Fetch /fetch/api/basic/../resources/redirect.py?location=../resources/top.txt with same-origin mode
PASS Fetch /fetch/api/basic/../resources/redirect.py?location=http://web-platform.test:8800/fetch/api/resources/top.txt with same-origin mode
PASS Fetch /fetch/api/basic/../resources/redirect.py?location=https://web-platform.test:9443/fetch/api/resources/top.txt with same-origin mode
PASS Fetch /fetch/api/basic/../resources/redirect.py?location=http://www1.web-platform.test:8800/fetch/api/resources/top.txt with same-origin mode
